Integrate firebase
Log in to firebase is now required.

The stories can be uploaded to firebase via the upload button. This
was necessary because of challenges with firebase-document's handling
of arrays: arrays within arrays, at least, very much confused it, and
it was uploading splice records instead of array data. Clearly, it was
not made for this. Manual upload was the simplest solution.